PatternEditor: enforce action in handleElements
The function does now check whether it should add or delete something.

In addition, adding new notes is now only possible if there is no other
note at the corresponding grid point.

Prior to this commit a note could be deleted by either
hovering and clicking it or by clicking its grid cell. The latter option
was kept to ensure a certain degree of backward compatibility with the
UX found in version 1.2.X. There the user could only toggle notes at
grid points and not directly interact with them.

But the above behavior proofed problematic once we allowed to
create/delete notes in the `NotePropertiesRuler`. There it would be just
too easy to accidentally delete notes. And disabling grid-based deletion
just in there would result in a more inconsistent UX. So, it has to go
away altogether.

+++ src/gui/src/PatternEditor/PatternEditor.cpp
@@ -1382,18 +1382,26 @@ void PatternEditor::validateSelection() {
 	}
 }
 
-void PatternEditor::handleElements( QInputEvent* ev, Editor::Action action ) {
+void PatternEditor::handleElements( QInputEvent* ev, Editor::Action action )
+{
+	if ( action == Editor::Action::None ) {
+		return;
+	}
+
 	// Retrieve the coordinates
 	GridPoint gridPoint;
-	if ( dynamic_cast<QMouseEvent*>(ev) != nullptr ) {
+	if ( dynamic_cast<QMouseEvent*>( ev ) != nullptr ) {
 		// Element added via mouse.
 		auto pEv = static_cast<MouseEvent*>( ev );
 
-		// If there are already notes at the provided point, delete them.
 		auto notesAtPoint = getElementsAtPoint(
-			pEv->position().toPoint(), getCursorMargin( ev ), true );
+			pEv->position().toPoint(), getCursorMargin( ev ), true
+		);
 		if ( notesAtPoint.size() > 0 ) {
-			deleteElements( notesAtPoint );
+			if ( action == Editor::Action::Delete ||
+				 action == Editor::Action::Toggle ) {
+				deleteElements( notesAtPoint );
+			}
 			return;
 		}
 
@@ -1403,7 +1411,7 @@ void PatternEditor::handleElements( QInputEvent* ev, Editor::Action action ) {
 			gridPoint.setRow( m_pPatternEditorPanel->getSelectedRowDB() );
 		}
 	}
-	else if ( dynamic_cast<QKeyEvent*>(ev) != nullptr ) {
+	else if ( dynamic_cast<QKeyEvent*>( ev ) != nullptr ) {
 		gridPoint.setColumn( m_pPatternEditorPanel->getCursorColumn() );
 		gridPoint.setRow( m_pPatternEditorPanel->getSelectedRowDB() );
 	}
@@ -1426,15 +1434,17 @@ void PatternEditor::handleElements( QInputEvent* ev, Editor::Action action ) {
 	float fYValue = std::nan( "" );
 	auto property = m_property;
 	if ( m_instance == Editor::Instance::NotePropertiesRuler &&
-		 dynamic_cast<QMouseEvent*>(ev) != nullptr ) {
-		fYValue = static_cast<NotePropertiesRuler*>(this)->eventToYValue(
-			dynamic_cast<QMouseEvent*>(ev) );
+		 dynamic_cast<QMouseEvent*>( ev ) != nullptr ) {
+		fYValue = static_cast<NotePropertiesRuler*>( this )->eventToYValue(
+			dynamic_cast<QMouseEvent*>( ev )
+		);
 	}
 
-	// Perform the action.
+	// When reaching this point, we only want to add new notes.
 	m_pPatternEditorPanel->addOrRemoveNotes(
-		gridPoint, nKey, nOctave, bNoteOff, fYValue, property, action,
-		Editor::ActionModifier::Playback );
+		gridPoint, nKey, nOctave, bNoteOff, fYValue, property,
+		Editor::Action::Add, Editor::ActionModifier::Playback
+	);
 }
 
 void PatternEditor::deleteElements(
